fix(injection-toolkit): resolve all clippy warnings and format workspace
Apply cargo fmt across 37 files and fix all clippy -D warnings:
- Replace manual modulo checks with .is_multiple_of()
- Add transmute type annotations for safety clarity
- Remove unnecessary casts and explicit auto-derefs
- Use c"" string literals instead of CStr::from_bytes_with_nul
- Collapse nested if statements, elide needless lifetimes
- Convert manual loops to iterator .find()
- Add #[allow(dead_code)] for planned API surface not yet wired up
- Add missing # Safety doc sections on unsafe fns
- Fix manual range contains with !(range).contains()

+++ packages/injection_toolkit/core/itk-video/src/scaler.rs
@@ -1,7 +1,7 @@
 //! Frame scaling and pixel format conversion.
 
 use crate::error::{VideoError, VideoResult};
-use crate::{DEFAULT_HEIGHT, DEFAULT_WIDTH, BYTES_PER_PIXEL};
+use crate::{BYTES_PER_PIXEL, DEFAULT_HEIGHT, DEFAULT_WIDTH};
 use ffmpeg_next::format::Pixel;
 use ffmpeg_next::software::scaling::{Context as SwsContext, Flags};
 use ffmpeg_next::util::frame::video::Video as VideoFrame;
@@ -109,7 +109,7 @@ impl FrameScaler {
         // Copy the frame data to our output buffer
         // RGBA frames have a single plane
         let data = self.output_frame.data(0);
-        let linesize = self.output_frame.stride(0) as usize;
+        let linesize = self.output_frame.stride(0);
         let row_bytes = (self.output_width as usize) * BYTES_PER_PIXEL;
 
         // Handle potential padding in frame rows
